int to real

+++ labs/lab03/lab-03-6.py
@@ -1,6 +1,6 @@
 yardLength = float(input())
 yardWidth = float(input())
 houseLength = float(input())
-houseWidth = int(input())
-wage = (yardLength * yardWidth - houseLength * houseWidth) * 2
+houseWidth = float(input())
+wage = yardLength * yardWidth - houseLength * houseWidth * 2
 print(wage)
